+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+Features:
+
+  - not supporting the `OodCore::Job::NodeRequest` object so it is removed
+
 ## 0.0.1 (2017-04-17)
 
 Initial release!